+++ packages/alfa-media/tsconfig.json
@@ -7,8 +7,13 @@
     "src/condition/index.ts",
     "src/condition/not.ts",
     "src/condition/or.ts",
+    "src/feature/comparison.ts",
     "src/feature/feature.ts",
+    "src/feature/height.ts",
     "src/feature/index.ts",
+    "src/feature/orientation.ts",
+    "src/feature/scripting.ts",
+    "src/feature/height.ts",
     "src/index.ts",
     "src/list.ts",
     "src/matchable.ts",
